Project authentication finished

+++ frontend/src/components/WorkoutDetails.js
@@ -1,14 +1,22 @@
 import { useWorkoutContext } from "../hooks/useWorkoutContext";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 // date-fns
 import { formatDistanceToNow } from 'date-fns';
 
 const WorkoutDetails = ({workout}) => {
     const { dispatch } = useWorkoutContext();
+    const { user } = useAuthContext();
 
     const handleClick = async () => {
+
+        if(!user) return
+
         const res = await fetch(`/api/workouts/${workout._id}`, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         });
         const json = await res.json();
